tasks and tests

+++ src/categories/categories.service.ts
@@ -1,34 +1,4 @@
-// import { Injectable } from '@nestjs/common';
-// import { v4 as uuidv4 } from 'uuid';
-// import { JsonDB } from 'node-json-db';
-// import { Config } from 'node-json-db/dist/lib/JsonDBConfig';
-// import { Category } from 'src/entitites/category.entity';
-
-// @Injectable()
-// export class CategoriesService {
-//   getCategoriesById(id: string): any {
-//       throw new Error('Method not implemented.');
-//   }
-//   private db = new JsonDB(new Config('myDataBase', true, false, '/'));
-
-//   getAll(): Promise<Category[]> {
-//     return this.db.getData('/categories');
-//   }
-
-//   create(category: Category): Promise<Category> {
-//     category.id = uuidv4();
-//     this.db.push('/categories[]', category);
-//     return Promise.resolve(category);
-//   }
-
-//   delete(id: string): Promise<void> {
-//     const index = this.db.getIndex('/categories', id);
-//     this.db.delete(`/categories[${index}]`);
-//     return Promise.resolve();
-//   }
-// }
-
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { v4 as uuidv4 } from 'uuid';
 import { JsonDB } from 'node-json-db';
 import { Config } from 'node-json-db/dist/lib/JsonDBConfig';
@@ -63,10 +33,20 @@ export class CategoriesService {
     }
   }
 
-  delete(id: string): Promise<void> {
+  async delete(id: string): Promise<void> {
+    // Retrieve tasks and then filter them
+    const tasksObject = await this.db.getData('/tasks');
+    const tasks = Object.values(tasksObject);
+    const tasksForCategory = tasks.filter((task: any) => task.categoryId === id);
+
+    if (tasksForCategory.length > 0) {
+      // Tasks found for the category, return a ConflictException
+      throw new ConflictException('Cannot delete category with associated tasks');
+    }
+
+    // No tasks found, proceed with category deletion
     const index = this.db.getIndex('/categories', id);
     this.db.delete(`/categories[${index}]`);
-    return Promise.resolve();
   }
 }
